requirements.txt
numpy==1.22.3
pandas==1.4.2
Pillow==9.4.0
requests==2.27.1
scikit_learn==1.2.1
streamlit==1.18.1
